simplified the quote code in redirs
