+++ includes/struct_def.h
@@ -24,7 +24,6 @@ typedef enum	e_token_type
 typedef struct 	s_token
 {
 	t_token_type	type;
-	//int				env_var;
 	char			quote;
 	char			*text;
