+++ srcs/parsing/tokens.c
@@ -1,5 +1,26 @@
 #include "minishell.h"
 
+int		get_nb_tokens(const char *s, char c)
+{
+	int			i;
+	int			nb;
+	char		quote;
+
+	i = 0;
+	nb = 0;
+	quote = 0;
+	while (s[i])
+	{
+		quote = check_quote(s[i], quote);
+		if (s[i] != c && (s[i + 1] == 0 || (s[i + 1] == c && !quote)))
+				nb++;
+		i++;
+	}
+	DEBUG(printf("\nnb_tokens : %i\n", nb);)
+	return (nb);
+}
+
+
 static t_token	*fill_tokens(const char *s, char c, int words, t_token *tokens)
 {
 	int		i;
